Use `tint.Err` for error logging, extract listen address logging attribute
